Console BattleShip
This was my final project for my Introduction to C++ I class.  We
didn't get into classes (that's the second semester course) so this is
pure functional programming.  It needs serious work: I plan on
refactoring it down a good bit and also start learning how to move the
functions to classes.